chromium should pretend that its chrome
WhatsApp Web only supports on Chrome not Chromium. If we do not change
the user agent WhatsApp prompts us to install a newer version of Chrome
and refuses to open the app.

This has to be done in the `appify` script, since chromium will not
change the user agent if an instance is already running.